move startBuild in build.ts
