+++ src/extension.ts
@@ -55,7 +55,7 @@ export async function activate(context: vscode.ExtensionContext) {
         vscode.commands.registerCommand('openshift.component.describe', (context) => execute(Component.describe, context)),
         vscode.commands.registerCommand('openshift.component.describe.palette', (context) => execute(Component.describe, context)),
         vscode.commands.registerCommand('openshift.component.create', (context) => execute(Component.create, context)),
-        vscode.commands.registerCommand('clusters.openshift.build.start', (context) => execute(Component.startBuild, context)),
+        vscode.commands.registerCommand('clusters.openshift.build.start', (context) => execute(Build.startBuild, context)),
         vscode.commands.registerCommand('clusters.openshift.build.showLog', (context) => execute(Build.showLog, context)),
         vscode.commands.registerCommand('openshift.component.createFromLocal', (context) => execute(Component.createFromLocal, context)),
         vscode.commands.registerCommand('openshift.component.createFromGit', (context) => execute(Component.createFromGit, context)),
