+++ src/k8s/build.ts
@@ -1,11 +1,17 @@
 import { window } from "vscode";
 import { OdoImpl, Odo } from "../odo";
+import { Progress } from "../util/progress";
 
 export class Command {
+
+    static startBuild(comp: string) {
+        return `oc start-build ${comp}`;
+    }
 
     static getBuild(build: string) {
         return `oc get build -l buildconfig=${build} -o json`;
     }
+
     static showLog(text: string, build: string) {
         return `oc logs ${text}${build}`;
     }
@@ -21,10 +27,39 @@ export class Command {
     static delete(build: String) {
         return `oc delete ${this.delete}`;
     }
+
+    static buildConfig() {
+        return `oc get buildConfig -o json`;
+    }
 }
 
 export class Build {
     protected static readonly odo: Odo = OdoImpl.Instance;
+
+    static async getBuild(text: string) {
+        const buildConfigName = [];
+        const buildConfigData = await Build.odo.execute(Command.buildConfig());
+        const buildConfigJson: JSON = JSON.parse(buildConfigData.stdout);
+        buildConfigJson['items'].forEach((key: any) => {
+            buildConfigName.push(key.metadata.name);
+        });
+        if (buildConfigName.length === 0) throw Error('You have no build available to start');
+        return await window.showQuickPick(buildConfigName, {placeHolder: text});
+    }
+
+    static async startBuild(context: any) {
+        let buildName: string;
+        if (context) {
+            buildName = context.id;
+        } else {
+            buildName = await Build.getBuild("Select the build to start");
+        }
+        if (!buildName) return null;
+        return Progress.execFunctionWithProgress(`Starting build`, async () => {
+            return Build.odo.execute(Command.startBuild(buildName));
+        }).then(() => `Build '${buildName}' successfully started`)
+        .catch((err) => Promise.reject(`Failed to start build with error '${err}'`));
+    }
 
     static showLog(context: any) {
         let buildName: string;
@@ -34,5 +69,21 @@ export class Build {
 
         }
         Build.odo.executeInTerminal(Command.showLog(buildName, '-build'));
+    }
+
+    static rebuild(context) {
+        let buildName: string;
+        if (context) {
+            buildName = context.impl.name;
+        }
+        Build.odo.executeInTerminal(Command.showLog(buildName, '-build'));
+    }
+
+    static followLog() {
+
+    }
+
+    static delete() {
+
     }
 }
