+++ src/openshift/component.ts
@@ -59,27 +59,6 @@ export class Component extends OpenShiftItem {
         return command.catch((err) => Promise.reject(`Failed to create Component with error '${err}'`));
     }
 
-    static async startBuild(context: any) {
-        let buildName: string;
-        if (context) {
-            buildName = context.id;
-        } else {
-            const buildConfigName = [];
-            const buildConfigData = await Component.odo.execute(Command.buildConfig());
-            const buildConfigJson: JSON = JSON.parse(buildConfigData.stdout);
-            buildConfigJson['items'].forEach((key: any) => {
-                buildConfigName.push(key.metadata.name);
-            });
-            if (buildConfigName.length === 0) throw Error('You have no build available to start');
-            buildName = await window.showQuickPick(buildConfigName, {placeHolder: "Select the build to start"});
-        }
-        if (!buildName) return null;
-        return Progress.execFunctionWithProgress(`Starting build`, async () => {
-            return Component.odo.execute(Command.startBuild(buildName));
-        }).then(() => `Build '${buildName}' successfully started`)
-        .catch((err) => Promise.reject(`Failed to start build with error '${err}'`));
-    }
-
     static async del(treeItem: OpenShiftObject): Promise<string> {
         const component = await Component.getOpenShiftCmdData(treeItem,
             "From which Project do you want to delete Component",
